Move LED constants, clarify confusing names, and add Javadocs for LED subsystem
