+++ src/main/java/frc/robot/Constants.java
@@ -69,35 +69,6 @@ public final class Constants {
     WOOD_BOT_TWO_2025
   }
 
-  
-  /**
-	 * @link https://www.revrobotics.com/content/docs/REV-11-1105-UM.pdf
-	 */
-  public static class LEDPatterns {
-    public static final double HOT_PINK = 0.57;
-    public static final double DARK_RED = 0.59;
-    public static final double RED = 0.61;
-    public static final double RED_ORANGE = 0.63;
-    public static final double ORANGE = 0.65;
-    public static final double GOLD = 0.67;
-    public static final double YELLOW = 0.69;
-    public static final double LAWN_GREEN = 0.71;
-    public static final double LIME_GREEN = 0.73;
-    public static final double DARK_GREEN = 0.75;
-    public static final double GREEN = 0.77;
-    public static final double BLUE_GREEN = 0.79;
-    public static final double AQUA = 0.81;
-    public static final double SKY_BLUE = 0.83;
-    public static final double DARK_BLUE = 0.85;
-    public static final double BLUE = 0.87;
-    public static final double BLUE_VIOLET = 0.89;
-    public static final double VIOLET = 0.91;
-    public static final double WHITE = 0.93;
-    public static final double GRAY = 0.95;
-    public static final double DARK_GRAY = 0.97;
-    public static final double BLACK = 0.99;
-  }
-
   /** Checks whether the correct robot is selected when deploying. */
   public static void main(String... args) {
     if (robotType == RobotType.SIM_BOT) {
